refactor: extract CustomSystemCommand as a separate module
